// src/api/nexuscore/model/listWikiPagesV1GuildsMeWikiGetParams.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Nexuscore
 * The backend for all Everthorn internal services
 * OpenAPI spec version: 1.0.0
 */
import type { ListWikiPagesV1GuildsMeWikiGetSortBy } from './listWikiPagesV1GuildsMeWikiGetSortBy.ts';
import type { ListWikiPagesV1GuildsMeWikiGetSortOrder } from './listWikiPagesV1GuildsMeWikiGetSortOrder.ts';

export type ListWikiPagesV1GuildsMeWikiGetParams = {
/**
 * Filter by published status
 */
published?: boolean | null;
/**
 * Filter by category
 */
category?: string | null;
/**
 * Filter by tags
 */
tags?: string[] | null;
/**
 * Fuzzy search by page title (summary and content comes later)
 */
search?: string | null;
/**
 * Sort by field
 */
sort_by?: ListWikiPagesV1GuildsMeWikiGetSortBy;
/**
 * Sort order
 */
sort_order?: ListWikiPagesV1GuildsMeWikiGetSortOrder;
/**
 * The page number to return
 */
page?: number | null;
/**
 * The number of items per page
 */
page_size?: number | null;
};